Move ButtonProcessor to main project
