+++ LegacySignController/test/test_ButtonProcessor/ButtonProcessorTests.cpp
@@ -10,8 +10,9 @@
 // Reference any needed cpp files directly to avoid pulling in
 // references that can't be resolved with the "fake" Arduino environment.
 #include "Utils\DebugUtils.cpp"
-#include <ButtonProcessor.h>
-#include "../TestHelpers/MockButton.cpp"
+#include "ButtonProcessor\ButtonProcessor.cpp"
+#include "ButtonProcessor\ButtonAction.cpp"
+#include "..\TestHelpers\MockButton.cpp"
 
 int lastCallerId(-1);
 String lastActionName("");
